Move test to tensorrt.py
